Fixed clippy warnings & errors

+++ src/nes/devices/addressing_modes.rs
@@ -19,10 +19,10 @@ impl Cpu6502 {
     
     /// Absolute addressing mode
     pub fn addr_ABS(&mut self, bus: &Bus) -> u8 {
-        let lo = self.read(bus, self.pc as u16) as u16;
+        let lo = self.read(bus, self.pc) as u16;
         self.pc = self.pc.wrapping_add(1);
         
-        let hi = self.read(bus, self.pc as u16) as u16;
+        let hi = self.read(bus, self.pc) as u16;
         self.pc = self.pc.wrapping_add(1);
         
         self.addr_abs = (hi << 8) | lo;
@@ -32,10 +32,10 @@ impl Cpu6502 {
     
     /// Absolute addressing mode with X offset
     pub fn addr_ABSx(&mut self, bus: &Bus) -> u8 {
-        let lo = self.read(bus, self.pc as u16) as u16;
+        let lo = self.read(bus, self.pc) as u16;
         self.pc = self.pc.wrapping_add(1);
         
-        let hi = self.read(bus, self.pc as u16) as u16;
+        let hi = self.read(bus, self.pc) as u16;
         self.pc = self.pc.wrapping_add(1);
         
         self.addr_abs = (hi << 8) | lo;
@@ -50,10 +50,10 @@ impl Cpu6502 {
     
     /// Absolute addressing mode with Y offset
     pub fn addr_ABSy(&mut self, bus: &Bus) -> u8 {
-        let lo = self.read(bus, self.pc as u16) as u16;
+        let lo = self.read(bus, self.pc) as u16;
         self.pc = self.pc.wrapping_add(1);
         
-        let hi = self.read(bus, self.pc as u16) as u16;
+        let hi = self.read(bus, self.pc) as u16;
         self.pc = self.pc.wrapping_add(1);
         
         self.addr_abs = (hi << 8) | lo;
@@ -68,7 +68,7 @@ impl Cpu6502 {
     
     /// Zero Page addressing mode
     pub fn addr_ZPG(&mut self, bus: &Bus) -> u8 {
-        self.addr_abs = self.read(bus, self.pc as u16) as u16;
+        self.addr_abs = self.read(bus, self.pc) as u16;
         self.pc = self.pc.wrapping_add(1);
         self.addr_abs &= 0x00FF;
         
@@ -77,7 +77,7 @@ impl Cpu6502 {
     
     /// Zero Page addressing mode with X offset
     pub fn addr_ZPGx(&mut self, bus: &Bus) -> u8 {
-        self.addr_abs = self.read(bus, self.pc as u16) as u16 + self.x as u16;
+        self.addr_abs = self.read(bus, self.pc) as u16 + self.x as u16;
         self.pc = self.pc.wrapping_add(1);
         self.addr_abs &= 0x00FF;
         
@@ -86,7 +86,7 @@ impl Cpu6502 {
     
     /// Zero Page addressing mode with Y offset
     pub fn addr_ZPGy(&mut self, bus: &Bus) -> u8 {
-        self.addr_abs = self.read(bus, self.pc as u16) as u16 + self.y as u16;
+        self.addr_abs = self.read(bus, self.pc) as u16 + self.y as u16;
         self.pc = self.pc.wrapping_add(1);
         self.addr_abs &= 0x00FF;
         
@@ -100,7 +100,7 @@ impl Cpu6502 {
     
     /// Relative addressing mode
     pub fn addr_REL(&mut self, bus: &Bus) -> u8 {
-        self.addr_rel = self.read(bus, self.pc as u16) as u16;
+        self.addr_rel = self.read(bus, self.pc) as u16;
         self.pc = self.pc.wrapping_add(1);
         
         if (self.addr_rel & 0x80) != 0 {
@@ -113,10 +113,10 @@ impl Cpu6502 {
     /// Indirect addressing mode
     /// (implements a hardware bug)
     pub fn addr_IND(&mut self, bus: &Bus) -> u8 {
-        let ptr_lo = self.read(bus, self.pc as u16) as u16;
+        let ptr_lo = self.read(bus, self.pc) as u16;
         self.pc = self.pc.wrapping_add(1);
         
-        let ptr_hi = self.read(bus, self.pc as u16) as u16;
+        let ptr_hi = self.read(bus, self.pc) as u16;
         self.pc = self.pc.wrapping_add(1);
         
         let ptr = (ptr_hi << 8) | ptr_lo;
@@ -132,7 +132,7 @@ impl Cpu6502 {
     
     /// Indirect addressing mode with X offset (zero page)
     pub fn addr_INDx(&mut self, bus: &Bus) -> u8 {
-        let t = self.read(bus, self.pc as u16) as u16;
+        let t = self.read(bus, self.pc) as u16;
         self.pc = self.pc.wrapping_add(1);
         
         let lo = self.read(bus, (t + self.x as u16) & 0x00FF) as u16;
@@ -145,7 +145,7 @@ impl Cpu6502 {
     
     /// Indirect addressing mode with Y offset (zero page)
     pub fn addr_INDy(&mut self, bus: &Bus) -> u8 {
-        let t = self.read(bus, self.pc as u16) as u16;
+        let t = self.read(bus, self.pc) as u16;
         self.pc = self.pc.wrapping_add(1);
         
         let lo = self.read(bus, t & 0x00FF) as u16;
